refactor(#INJI-42): [Kiruthika/Bhargavi] extract vc item contents & activation status into separate components
Issue Link: https://mosip.atlassian.net/browse/INJI-42
